Create README.md
TATA-Data-visualization-Internship
Certificate

Task - 1
Framing the Business Scenario:
To prepare for your meeting, you need to draft questions that you think will be important and relevant to the CEO and CMO. This preparation will be your guide as you develop your presentation.

For this task, you are only required to draft the questions. Make sure to think both quantitatively and qualitatively.

You’ve been provided a dataset in the resources below to use as the basis for your exploration. Review this data, taking note of what information has been provided, what insights you can garner, and what is relevant to both the CEO and CMO respectively.

So, prepare 8 questions(4 for the Chief Executive Officer & 4 for the Chief Marketing Officer)

Answer : Task 1

Task - 2
Choosing the Rights Visual:
In this task, you will be required to read the questions carefully and understand that business requirement. Once you have an idea of what is required from the perspective of the CEO and CMO, you will need to come up with the perfect visual which will illustrate what the senior managers are looking for in each scenario. Remember, data can be presented in multiple types of charts, but you are required to select the visual that would best display the information which is being presented.

You will be provided resources on how to select visuals based on the different scenarios, these are available in the resources section. These resources will help you get an idea on which visual to select for the given business scenario and will also guide you on how to choose the right chart or graph for your data. Each question will contain a unique scenario and you will be expected to answer the questions based on that scenario.

Answer : Task 2

Chart Guide:
Comparison:
Column chart
Bar chart
Line chart
Scatter plot
Distribution(Spread of data values over categorised and continous values):
Histogram
Box plot
KDE plot
Breakup of a whole chart(various parts comprise the whole):
Pie chart
Donut chart
Stacked Column chart
Stacked Bar chart
Relationship charts(relationship between two variables):
Scatter plot
Line chart
Trend charts(trends of values over times and categories/also known as time series data):
Line chart
Area chart
Column chart
Chart Guide:


Task - 3
Creating Effective Visuals:
Data Cleanup:

Create a check that the quantity should not be below 1 unit
Create a check that the Unit price should not be below $0
Revenue = 'Online Retail'[Quantity]*'Online Retail'[UnitPrice]

Create each visual on a separate tab

3.1
The CEO of the retail store is interested to view the time series of the revenue data for the year 2011 only. He would like to view granular data by looking into revenue for each month. The CEO is interested in viewing the seasonal trends and wants to dig deeper into why these trends occur. This analysis will be helpful for the CEO to forecast for the next year.

Answer : 

3.2
The CMO is interested in viewing the top 10 countries which are generating the highest revenue. Additionally, the CMO is also interested in viewing the quantity sold along with the revenue generated. The CMO does not want to have the United Kingdom in this visual.

Answer : 

3.3
The CMO of the online retail store wants to view the information on the top 10 customers by revenue. He is interested in a visual that shows the greatest revenue generating customer at the start and gradually declines to the lower revenue generating customers. The CMO wants to target the higher revenue generating customers and ensure that they remain satisfied with their products.

Answer : 

3.4
The CEO is looking to gain insights on the demand for their products. He wants to look at all countries and see which regions have the greatest demand for their products. Once the CEO gets an idea of the regions that have high demand, he will initiate an expansion strategy which will allow the company to target these areas and generate more business from these regions. He wants to view the entire data on a single view without the need to scroll or hover over the data points to identify the demand. There is no need to show data for the United Kingdom as the CEO is more interested in viewing the countries that have expansion opportunities.

Answer : 

Task - 4
Communicating Insights and Analysis
Develop a script and record a video presenting your findings to the CEO and CMO based on the four questions they asked and the visuals you created in the previous tasks.

Answer : Task 4

Note :This video isn't made by me

About
TATA Data Visualization Internship

Topics
data-visualization data-cleaning
Resources
 Readme
 Activity
Stars
 0 stars
Watchers
 1 watching
Forks
 1 fork
Report repository
Releases
No releases published
Packages
No packages published
Footer
© 2025 G